Modernise modalShim
Do them as I have to work on them I guess. We'll get there in the end.
Contributes to #1804
